refactor : IDE-8 IDEPage 내 스타일 파일 컴포넌트 별 분리
