+++ src/pages/IDEPage/Header/IDEHeader.tsx
@@ -1,6 +1,6 @@
-import { IDEHeaderContainer, IDELogo } from './IDEPage.style';
-import StyledLink from '../../styles/StyledLink';
+import StyledLink from '../../../styles/StyledLink';
 import CurrentUser from './CurrentUser';
+import { IDEHeaderContainer, IDELogo } from './IDEHeader.style';
 
 const IDEHeader = () => {
   return (
