+++ src/pages/IDEPage/Terminal/Terminal.style.ts
@@ -1,0 +1,12 @@
+import { styled } from 'styled-components';
+
+export const TerminalContainer = styled.div`
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  position: fixed;
+  bottom: 0;
+  height: 300px;
+  width: 100%;
+  background: #d9d9d9;
+`;
